product/automotive-rd: Introduce gicx00_multiview module
Certain GICs have a "multiview" (multiple view) feature, allowing
firmware to allocate GIC resources into up to three views. This allows
systems to allocate the GIC to up to three different OS or hypervisors
that are running independent software stacks.

Introduce the gicx00_multiview module (with documentation) to configure
one or more multiview GICs by assigning individual redistributors to
views then assigning individual interrupts to views.

If SCP-firmware is using one of the configured multiview views then the
gicx00_multiview module must be listed prior to gicx00 in SCP_MODULES so
that it is initialized first.

It checks whether multiview support is enabled on the hardware variant.
Initialization proceeds only if it is supported; otherwise, it returns
immediately without raising a hard fault.

Upstream-Status: Pending
Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>
Signed-off-by: Divin Raj <divin.raj@arm.com>