prod/rdaspen: Configure PFDI monitor modules
Add the required configurations for the PFDI monitor and the
SCMI PFDI monitor modules in addition to the required timer,
MHUv3, transport and SCMI configurations.

Also align the value of SI0_SCMI_PAYLOAD_SIZE with the value
used in Trusted Firmware-A so the memory mapping is aligned
on both sides.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>