prod/rdaspen: Increase shared SRAM size
Increase the size of shared SRAM between the primary compute
and Safety Island to 8KB to accommodate the memory required for
the SCMI PFDI monitor communication where each core requires
at least 32 bytes.

Note: The extra 4KB are not fully utilized at the moment, but the
region needs to be 4KB aligned to meet memory mapping constraints.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>