prod/rdaspen: Handle PD transitions in PFDI Monitor
Currently, if a CPU goes off, it would stop sending
PFDI Monitor messages, resulting in the failure of
PFDI monitoring.

Subscribe to the power domain transitions in the
PFDI monitor module so that it stops PFDI monitoring
if the CPU goes off and turns on the PFDI monitoring
again if the CPU goes on.

Also increase the maximum number of allowed notifications
so that it becomes enough.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>