mod/cmn_cyprus: Force non-hash region address
Force the region size calculation to use the start address directly
(masked) without dividing by the min size. (Not sure why this isn't
working as expected)

Also apply a bit-mask to the end address as the lower bits are
meaningful in regions 2-63.

TODO: Figure out how this calculation is meant to work. Maybe this never
worked as all the example configurations have a start address of zero
for the non-hash regions?

Upstream-Status: Inappropriate [Proper upstream solution required]
Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>