mod/scmi_perf: Stop SCMI perf FCH events on debug CLI entry
After entering the debug CLI, the fast channel pending
requests are not processed anymore. Due to this,
the event buffer is filled up resulting in a crash.

Stop the SCMI fast channel (FCH) events to be put in the
queue when the CLI is active. Enable back when CLI is
exited.

Upstream-Status: Inappropriate [Workaround until FCH event queue fix]

Signed-off-by: Archish Venkatesh <archish.venkatesh@arm.com>