prod/automoive-rd: GIC multiview module improvements
When the GIC multiview module configures an ATU-mapped GIC across
the CMN, performing this setup from the element_init function fails
because the CMN is not yet initialized (its configuration happens
in the CMN start function).

However, the Safety Island’s GIC itself is initialized in the GIC
module’s init function.

To address this, introduce a delayed flag that defers multiview
configuration until the CMN start phase.

Additionally, remove the redistributor wake-up step, as this is
handled later by the GIC module itself.

Also fix the doxygen comments in the header file of the module
by removing the brief prefix as it is not appropriate to use
for the elements of structures and enums.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>