prod/rdaspen: Remove MHU sync design for RAS sync
This removes the MHU sync between the AP and SI0
the current implementation on SI0 side only covers
the following Outband errors.

Which covers the following error type for CPU IP:
- UE - Uncontainable Errors

The AP will assert and not be able to return to
normal execution, but SI will clear the error and
raise the SSU Alarm.

This updates the GIC multiview with the new ERR
interrupts instead of FHI interrupts

Upstream-Status: Pending
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>